Moves env completer to comp
